Fixed additional logic errors in the alljoyn_c BusAttachmentTest
Multiple tests were trying to add a interface to a BusObject after the bus object
had been registered with the BusAttachment.  This caused the tests to fail with a
ER_BUS_CANNOT_ADD_INTERFACE error.  By adding the interface to the BusObject
before registering the BusObject this error can be avoided.