Updated the samples so they reflect recent changes
Callback functions call enableconcurrentcallbacks when they calling another
alljoyn method.

interface name, well-known name, and object path have been updated to they match
the name used in the C++ samples.

The call to connect no longer tries all of the connect specs it calls connect
with the Linux connect spec.  The default behavior is to default to the null:
transport if it is unable to connect.  For this reason it will always pass or
fail on the first call.

The Security sample calls alljoyn_proxybusobject_enablepeersecurity so the user
has more than 5 seconds to enter the password for the security.

Change-Id: I36cb62f0ac7f1132062b443c2bc4963c8c15c08a