ALLJOYN-1712 Fixed Memory leak due to usage of the DeferredCallback class
The DeferredCallback class is in the C bindings for one reason to force a
Callback function to return on the same thread that registered the callback.
The DeferredCallback class is only important for the Unity Extension on Android
phones.  This is because on Android phones Unity uses a single-threaded
apartment model. The callback must return on the same thread that the listener
was registered on.

The DeferredCallback class is only used if the user manually calls the
alljoyn_unity_set_deferred_callback_mainthread_only function.  If the user did
not call this function the DeferredCallback class was still allocated and used
however it was not deleted.

These changes will no allocate a DeferredCallback unless a user has called the
mainthead only function.  I have also modified the DeferredCallback class its
self so it will not add the callback to the callback queue unless the mainthread
function has been called.  This was probably not needed since with the other
changes this should never be called.  It is a preventive measure to make
sure callbacks placed in the sPendingCallbacks list will be removed when the
callbacks are processed by the DeferredCallback::TriggerCallbacks call.

Also added the unit test BusListenerMainThreadTest this runs the exact same
tests that are found in BusListenerTest except that it is run in maintread_only
mode.  This ensures that both modes are tested by the unit tests.

Changed (int) to QCC_BOOL in the
alljoyn_unity_set_deferred_callback_mainthread_only function since it was a
boolean value.

Changed allJoyn_typeid to alljoyn_typeid it should have all been lower case.

Fixed a few compiler errors identified when compiling using GCC that were not
discovered when compiling using MSVC.

Updated the .gitignore to exclude files copied in by SCons and files generated
by doxygen.

Change-Id: Id204019648d93091377a0a617a5fe2ef4c0be58c