ALLJOYN-1730 Accessor functions for alljoyn_sessionopts
Befor this change the only accessor functions that existed for the
alljoyn_sessionopts was a get function.  There was no way to change a value
in an existing sessionopts.  This meant that the only way to change the
sessionopts was to destroy the old sessionopts and create a new one.  With
this change that is no longer needed.

Change-Id: I92e19c20badd70c775c02b7a83d3ddfc3197cb8c