ALLJOYN-1676 Changed the way values are assigned to library env variables
The scripts used the fact that the libraries were always in the same location
when in the dlibs variable.  With the change for ALLJOYN-1676 this was no longer
always true.  So we must search list to find the scons.file.nodes that contain
the shared and static alljoyn_c libraries.