ALLJOYN-1710 subtask ALLJOYN-1798 Implement setlinktimeout in C
This code includes basic unit tests for the
alljoyn_busattachment_setlinktimeout and the
alljoyn_busattachment_setlinktimeoutasync functions.
The unit test only test that the bindings actually call all of the
way through to the underlying C++ methods.  It does not test that
when setlinktimeout is set that the daemon will actually drop an
unresponsive session.

Change-Id: I8369dfda945de60cea3511bb495061436200cb42