ALLJOYN-1710, subtask ALLJOYN-1791 Add Async AuthListeners to C binding
The alljoyn_authlistener is now passed as one of the parameters for the
alljoyn_authlistener_requestcredentialsasync_ptr and the
alljoyn_authlistener_verifycredentialsasync_ptr.  The reason for this is when
using the async authlisteners the user is required to always call the
corresponding response functions that require the alljoyn_authlistener as one
of the parameters. If this parameter were not passed in the user would have to
pass in the alljoyn_authlistener as part of the context or make it a global
value.

Fixed memory leaks and compiler warnings seen when using the gcc compiler and
valgrind.

Change-Id: I43b358f6353372a3e7ecf1a8611cf3395bcc74bf