Updated the binding for the alljoyn_keystorelistener added unit tests.
A new parameter has been added to the loadrequest and storerequest callbacks.
The listener parameter.  Since the user is required to call the
alljoyn_keystorelistener_putkeys and alljoyn_keystorelistener_getkeys from the
listeners.  The need to have the alljoyn_keystorelistener to pass into those
commands as the first parameter.

Added the KeysStoreListenerTest to the unit test.  Updated the AuthListenerAsyncTest
so that its callback functions were static.  Without making them static the the
AuthListener callback functions were global and caused conflicts with other callback
functions with the same name in the KeyStoreListenerTest.

Change-Id: I52ce3245a13124d8884efeba17b95d5197faf028