Changed the name of the static library alljoyn_c.lib to alljoyn_c_static.lib
When generating the shared library an alljoyn_c.lib is generated that contains
the Import libs.  This file is used for linking with the .dll file.  However
when the static library a alljoyn_c.lib file was also created that contained the
standard static libraries.  We just got lucky that our scons script always built
the static library after the dynamic library.  In some situations when building
with a -j options the Import lib would be used when we wanted the static lib to be
used.

By changing the name of the static library there is no question which .lib file
should be used.

Both versions of the file are installed in the dist/lib folder.

Change-Id: I1cf2a24067e6a4bc5171cb03839c6e8045567257