Only name the static library alljoyn_c_static on Windows
Changing the name of the static library is only needed on Windows.
Not naming the file with a different name on posix based computers
makes is simpler to link against the library since the only change
needed to make is a static link library is to add the -s liker option

To understand the reason the file is renamed alljoyn_c_static in Windows
please see the comments in the src/SConscript file.

Change-Id: Idfe54f159ac77f94fd057e036f03206bc5ea23c4